Refactor ssh key generation function to a new ssh utils file
